requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
discord.py==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mutagen==1.46.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
websockets==11.0.3
yarl==1.9.2
yt-dlp==2023.3.4
